requirements.txt
pyvisa>=1.13
numpy>=1.24
matplotlib>=3.7
pyyaml>=6.0
